refactor: move ``Emoji`` to own module (#978)
* refactor: separate Emoji to own module

* fix imports

* rename imports in http

* change brackets

* ci: pre-commit

* fix: fix incorrect role converts in Emoji

* ref: remove Role import
