+++ interactions/api/http/emoji.py
@@ -1,8 +1,9 @@
 from typing import List, Optional
 
 from ...api.cache import Cache
-from ...api.models.guild import Emoji, Guild
-from ...api.models.misc import Snowflake
+from ..models.emoji import Emoji
+from ..models.guild import Guild
+from ..models.misc import Snowflake
 from .request import _Request
 from .route import Route
