+++ interactions/api/http/invite.py
@@ -4,7 +4,7 @@ from ...api.cache import Cache
 from .request import _Request
 from .route import Route
 
-__all__ = ["InviteRequest"]
+__all__ = ("InviteRequest",)
 
 
 class InviteRequest:
