+++ interactions/client/get.py
@@ -14,9 +14,10 @@ from sys import version_info
 
 from ..api.error import LibraryException
 from ..api.http.client import HTTPClient
+from ..api.models.emoji import Emoji
 from ..api.models.guild import Guild
 from ..api.models.member import Member
-from ..api.models.message import Emoji, Message
+from ..api.models.message import Message
 from ..api.models.misc import Snowflake
 from ..api.models.role import Role
 from .bot import Client
